greenlight.conf: Explicitly specify the blindsidenetworks test server
The documentation claims that if the BBB credentials are not specified
it will automatically use the blindsidenetworks.com test server, but
that appears not to be the case. So we specify the default credentials
found at:

    https://github.com/bigbluebutton/greenlight/blob/master/config/application.rb

The connectivity test to that test server still isn't working, failing
with:

    Checking environment: Passed
    Checking Connection: Passed
    Checking Secret: Failed
    Could not get a valid response from BigBlueButton server - <html>
    <head><title>504 Gateway Time-out</title></head>
    ...

It appears that they have a proxy in front of the test server that's
working, but the test server itself, behind the proxy, is down.