greenlight.nginx: Config file extracted from greenlight Docker image
From step 4:

    docker run --rm bigbluebutton/greenlight:v2 cat ./greenlight.nginx

It mentions that one can use a relative root other than /b using
instructions here:

    https://docs.bigbluebutton.org/greenlight/gl-config.html#using-a-different-relative-root

This also discusses SQLite vs. PostgreSQL in the Greenlight container;
we will probably want to mount the database files from a source
external to the container for backup and easy movement of the server.